refactor: 블랙잭 덱 생성 테스트 - 카드를 확인하도록 수정
