+++ src/main/java/domain/strategy/DeckGenerateStrategy.java
@@ -1,0 +1,8 @@
+package domain.strategy;
+
+import domain.blackjackgame.BlackjackDeck;
+
+public interface DeckGenerateStrategy {
+
+    BlackjackDeck generateDeck(DrawStrategy drawStrategy);
+}
