+++ src/main/java/domain/strategy/DeckGenerator.java
@@ -1,0 +1,10 @@
+package domain.strategy;
+
+import domain.blackjackgame.BlackjackDeck;
+
+public class DeckGenerator {
+
+    public BlackjackDeck generateDeck(DrawStrategy drawStrategy, DeckGenerateStrategy deckGenerateStrategy) {
+        return deckGenerateStrategy.generateDeck(drawStrategy);
+    }
+}
